Move HostThemeManager and HostThemeWatcher under theme extensions directory
