+++ packages/app/src/cli/utilities/extensions/theme/host-theme-watcher.ts
